requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.2
python-dotenv==1.0.1
openai==1.70.0
google-cloud-storage==3.1.0
python-telegram-bot==22.0
yt-dlp==2025.3.31
